Remove owner_id from schema and fix typo in INSERT statement
When the cats table was created, it had the following columns: id, name, age, and breed. However, in the "Creating the Table" section, the lesson incorrectly shows that running .schema would return a cats table containing an owner_id column. As for the typo, Hana's age was incorrectly set to 3, when the lesson showed earlier that her age is 1.